Fix timezone extraction to work in all locales
Followup to #21585
