+++ test/unistd/sysconf64.out
@@ -118,7 +118,7 @@ _SC_CLK_TCK: 100
 _SC_HOST_NAME_MAX: 255
 _SC_AIO_PRIO_DELTA_MAX: 0
 _SC_STREAM_MAX: -1
-_SC_TZNAME_MAX: 6
+_SC_TZNAME_MAX: 16
 _SC_THREAD_DESTRUCTOR_ITERATIONS: 4
 (invalid): -1
 errno: Invalid argument
